+++ analytics/upress_reporting/models/data/download_event.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+@dataclass(init=True, repr=True)
+class DownloadEvent():
+    title: str
+    timestamp: str
+    edition_id: int